backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
sqlalchemy
psycopg2-binary
joblib
python-dotenv
passlib[bcrypt]
bcrypt
PyJWT
pandas
scikit-learn
joblib